Consider this the combined commit message for this and r835.
Replaced the venue dropdown on the event form with an autocompleting text field. If the venue is not found, it will be created and the user will be forwarded to a form it edit it.
Added scoped searching to DuplicateChecking#find_with_duplicate_support so that it will combine existing :conditions with the duplicate checking query.
Added yeilded content_for blocks to the common header so that individual pages can insert javascript or css easily.



git-svn-id: https://calagator.googlecode.com/svn/trunk@836 54c6411d-1744-0410-a9eb-b99409f43359