Fixed sorting of tag search results. Sorting by date, event title, and venue title are supported.
To answer Igal's question in the (now removed) FIXME: The call to add_joins was allowing us to add associations into the tagged_with SQL query. It was throwing an exception because it was being called with the entire options hash instead of just the specific option that it wanted to see.