Fixed Event::to_ical: added calendar name and tests.
The Event::to_ical method adds the name of the calendar. It relied on a specific string and letter case that worked with the old VPIM exporter, but not with the new RiCal exporter. This injector's been fixed and improved to eliminate these assumptions, and examples have been added to describe the intended behavior of the iCalendar headers.