Refactored SearchEngine::Sunspot to return more meaningful results.
In the past, a single giant query was done to fetch all results, but this was
bad because it would return a subset of a globally sorted set, so that sorting
by relevance might produce a result set without any current events. Next
approach tried was to retrieve the most recent matches and sort them using
Ruby, which was convoluted. Finally, this approach performs two separate
searches on current and past events, and by default will show the most relevant
matching current events and the most recent matching past events.