Fixed SearchEngine::Sunspot so `rake spec` works w/out Sunspot installed.
The SearchEngine::Sunspot driver class was originally written with the 
assumption that it would only ever be loaded if Sunspot was selected as the
search engine. 

Unfortunately, the Rails test environment loads all classes that in the 
autoload paths it finds, while the development and test environments only
load those they need. This was a problem because if a person doesn't have
Sunspot installed and configured, this class would be loaded and cause the
execution of the specs to fail.

The fix is to only monkey patch Sunspot if it's present so this class can be
loaded safely even if Sunspot isn't being used.