HACK: Hack to send garbled mail on exception notifications.
The combination of Rails 2.3.10, theme_support and exception_notifier
doesn't work. The theme_support plugin patches how template paths work
in such a way that the exception_notifier plugin can't find its templates.
With this workaround, the plugin can at least find its plugin.
Unfortunately, the content of the email is garbled such that text runs
together, e..g. arrays are printed with #to_s rather than #inspect. The
value of this workaround is that the email is at least sent so that the
admin can go and check the logs to figure out what's going on.

The exception notification system needs to be replaced with something else
that actually works. The upgraded version of this plugin, at version 3.0.18,
doesn't work any better and emits the same garbled text. Ugh.