Improve venues#destroy to prevent removal of venues with events.
When someone deleted a venue with events, they'd all lose their venue
associations and look weird. Restoring the deleted venue wouldn't bring
back the associations because the `:dependent => :nullify` bypasses the
versioning system.

The new behavior is to prevent deletion of a venue that has events and
display a message telling them to reassign all the events first. This
should make it harder to destroy a venue that's in use.