THEME: Remove "tz" setting, only "timezone" is needed now.
The Rails environment has a `config.time_zone` to specify the timezone for
ActiveRecord date-time conversions, but the rest of the app doesn't
know about this and instead uses the OS user's "TZ" timezone.

In the past, theme authors had to specify both the Rails and OS timezones
separately, which was annoying because these have different formats.

The solution is to let the "timezone" setting tell Rails what the current
timezone is, then figure out what OS timezone this maps to and set to that.