Add Plancast and Meetup parsers using new wrapper mechanism.
A new "to_abstract_events_wrapper" is introduced which provides a way to
hand off parsing to another driver.

For example, the Plancast driver is
just says that it's a wrapper around the Ical driver, and that when someone
enters a Plancast URL, that a different URL should be retrieved and handed
off to the Ical parser.