Fix Google Calendar export by truncating more vigorously and before escaping.
Export to Google Calendar is a pain because there's no published API and
the only known way is to create a giant GET parameter. Unfortunately,
description strings are long and can overrun the maximum length of a
permitted URL.

Worse still, Google randomly does redirects that add hundreds of additional
characters to the GET string. This patch significantly reduces the length
of the description included to cope with this.

Also, the earlier code was escaping and then truncating, which could
produce invalid URI escape sequences. Now the text is truncated first
and then escaped.