Disabled default theme's unused theme-specific JavaScript because "theme_support" plugin has problems.
We never used the theme-specific JavaScript because this feature never
worked. The "theme_support" plugin can't serve the contents of this file
because of low-level bugs in its implementation, and if this wasn't enough,
the path it generates for the "theme_javascript_include_tag" is invalid.
Fixing the "theme_support" plugin to do this correct requires many
changes throughout the code and would be quite invasive.