Revert "Remove need for a Google geocoder key."
This reverts commit 9e3df9d6d540cc9a0ffbbd5e73c225a31593940e.

We were under the mistaken impression that no Google Maps key was needed anymore because (1) the Google Maps geocoder no longer seemed to need a key and (2) the Google Maps worked fine on "localhost" with the key set to "Google".

Unfortunately, this was an incorrect conclusion because if the server is accessed as anything other than "localhost", an obnoxious error popup will be displayed instead of a map.

Therefore, these changes are being reverted. Sorry.