Fix Gemfile by making "ruby-debug" a dependency.
Previously I used trickery to load "ruby-debug" using #require in hopes
of only loading it when it was available. Unfortunately, this caused it
to load the latest versions of a few gems, which could be at odds with
what the Gemfile specified, such as "linecache" and "columnize".

The solution is to make "ruby-debug" a real dependency.