Fix venue_spec to destroy existing versions before counting them.
There's a bug somewhere in the stack that causes database sequences to be
reused during tests. For example, one spec creates a Venue with id 123 and
then deletes it, but the version information stays behind. Later, another
Venue is created and is also given id 123. This causes the "paper_trail"
versioning system to think it's related to the previously deleted record,
rather than treating it as an unrelated record.