Fix "events" fixtures to use proper times.
The problem is that Rails 3 or RSpec 2 relaxed how dates are parsed
in fixtures, and SQLite was too permissive and allowed non-dates to
be inserted into columns. This resulted in confusing behavior, because
strings are always after dates, e.g. the date 2014-12-31 is before the
string "2010-01-01".

The solution was to rework the generated times in the fixtures so that
they are all properly specified.