Fix Upcoming parser to cope with new API's invalid UTC dates.
The Upcoming API now emits "utc_start" and "utc_end" dates that have
nothing to do with the event, these can be off by months. I suspect that
these dates are coming from the first event in a recurring series.

The solution is to use data from fields like "start_date" and "start_time",
which used to be invalid, but are now the only way to get this information.
Hopefully they won't screw those up again.