Fix events#_form to handle invalid dates sanely and without duplicated code.
Previously sensible defaults were only provided if the event was a new
record, now they're always set if needed.

Previously there was lots of duplicated code handling invalid dates which
didn't catch some cases, such as an existing record with an invalid date;
now this is fixed and the code is clearer thanks to default behavior.