Fix #updated_at, which may have been breaking iCal refreshes.
Some iCalendar clients rely on the timestamps provided by the #updated_at
colum, which wasn't being updated, so they assumed the record hadn't
changed despite the SEQUENCE number and content changing.

The problem is that Rails 3 does not update the #updated_at column on
#save, only on #create and #update_attributes. That's really stupid.

This change reimplements the old behavior of updating the #updated_at
column on #save as a mixin.