Asset pipeline support, small theme change required [THEME] [DEPLOY]
It should now be possible to develop themes which take advantage of the
asset pipeline for non-view resources such as stylesheets using SASS/SCSS
or javascript using CoffeeScript (add to :assets group in Gemfile).

[THEME] If you are maintaining a theme or fork of calagator, you will
need to make a small change to your templates, due to incompatibilities
between the asset pipeline and the ':cache => true' attribute.

  1. Change following line in themes/*/views/layouts/application.html.erb:
       <%= javascript_include_tag  :defaults, :cache => true %>
     to look like:
       <%= javascript_include_tag 'application' %>

  2. Remove any additional ':cache => true' attributes used in your
     theme; most likely you won't find any beyond #1 above.

[DEPLOY] Deploying calagator now requires you to precompile assets using:

  $> rake assets:precompile