Refactor url validator in Event model
The regex used by the url validator was allowing invalid urls to be saved.
Urls of the form `example.com, http://www.example.com` were allowed to be
persisted to the database. Added a ^ to the regex to start it at the beginning
of the line, and simplified how it checked for the http/https scheme.