Provide an example, but don’t set default_url_host in secrets.yml.sample
Setting this in the sample causes test failures due to expectations
written using explicit http://test.host URLs. I looked into adding specs
for this feature directly, but this would involve stubbing things on the
global SECRETS object and will be moot once #314 is merged and we can 
move this to being a direct configuration on the containing app.